refactor: move Wheel component logic to WheelGameControls and remove redundant file #155
